Revert "[Feature] 꽃 주고 받기 통신 코드 및 기본 UI 작업"
