+++ liveOnReboot/liveOn/Network/Gift/Flower/DTO/FlowerDTO.swift
@@ -4,14 +4,17 @@
 //
 //  Created by Keum MinSeok on 2022/07/09.
 //
+
 import SwiftUI
 
-struct FlowerGetResponse: Codable {
-    let flowerType: String
-    let message: String
+struct FlowerDTO: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
 }
 
-struct FlowerPostRequest: Codable {
-    let flowerType: String
-    let message: String
+struct FlowerDTO_Previews: PreviewProvider {
+    static var previews: some View {
+        FlowerDTO()
+    }
 }
