+++ liveOnReboot/liveOn/View/SendGift/SendPicture/SendPictureView.swift
@@ -85,6 +85,7 @@ struct SendPictureView: View {
 
                                     imagePost {
                                         print("전송완료!")
+                                        showLoading.toggle()
                                         isSent.toggle()
                                     }
                                 })
